[inductor] Address yushangdi review: lowering_fx_node, move get_fused_kernel_module_fqn, unit tests
Address seven review comments from yushangdi on PR #187061:

- Use torch.cuda.graph(g, enable_annotations=True) instead of calling
  enable_annotations() / resolve_pending_annotations() directly in the
  test and docstring example.  These two are internal to the context
  manager when enable_annotations=True is passed.

- Replace origin_node (repurposed from its original None value in
  IRNode.__post_init__) with a new lowering_fx_node field that is
  populated from V.get_current_node() during lowering.  This avoids
  touching origin_node semantics and reuses the existing
  V.set_current_node(n) context instead of adding a separate
  current_primary_node context manager.

- Move get_fused_kernel_module_fqn from graph_view.py to scheduler.py.
  The function operates on scheduler nodes and is called from
  scheduler.py and simd.py, so it belongs there rather than in the
  FX-pass graph view module.

- Access V.graph._current_kernel_module_fqn directly in wrapper.py
  (it is a declared attribute) rather than via getattr with a fallback.

- Add unit tests for _clean_stack_name and _strip_instance_suffix in
  TestGraphViewHelpers.

Test Plan:

```
python -m pytest test/inductor/test_cudagraph_fqn_annotations.py -v
```
