+++ torch/_inductor/codegen/simd.py
@@ -4034,7 +4034,7 @@ class SIMDScheduling(BaseScheduling):
             coalesce_analysis = None
 
         if torch._inductor.config.triton.cudagraph_kernel_annotations:
-            from torch._inductor.fx_passes.graph_view import get_fused_kernel_module_fqn
+            from torch._inductor.scheduler import get_fused_kernel_module_fqn
 
             V.graph._current_kernel_module_fqn = get_fused_kernel_module_fqn(nodes)
         else:
