+++ torch/_inductor/fx_passes/graph_view.py
@@ -6,7 +6,6 @@ import re
 from typing import Any, TYPE_CHECKING
 
 import torch.fx as fx  # noqa: TC001
-from torch._inductor.virtualized import V
 from torch.utils._ordered_set import OrderedSet
 
 
@@ -122,96 +121,6 @@ def _strip_instance_suffix(name: str) -> str:
 def _outermost_prefix(stack: Any) -> str:
     """Return the cleaned outermost (block-level) module path from an nn_module_stack."""
     return _clean_stack_name(next(iter(stack.values()))[0])
-
-
-def get_fused_kernel_module_fqn(scheduler_nodes: Any) -> str | None:
-    """
-    Return a human-readable FQN annotation for a fused kernel.
-
-    Uses V.graph.fx_fqn_map — built once during lowering in graph.py:run_node —
-    to map FX node names to FQN strings, e.g.
-    "convolution_1" -> "L.networks.1.conv.convolution".
-
-    Two-pass hybrid algorithm:
-
-    Pass 1 — anchor prefixes via origin_node:
-        For each snode, origin_node is the direct FX node whose run_node
-        call created the IR buffer.  It is never transitively accumulated,
-        so it gives an unambiguous block identity.  Look it up in fqn_map
-        to get the anchor FQN, then read its nn_module_stack outermost
-        entry to get the block-level prefix (e.g. "L.networks.3").
-
-    Pass 2 — collect FQNs from origins with prefix filter:
-        Walk all origins across every snode.  origins is transitively
-        accumulated (cascading history from upstream blocks), but each
-        entry is looked up in fqn_map and filtered by the anchor prefix.
-        This correctly captures inline ops (relu, add) that share the
-        same block but were inlined into a parent buffer and never became
-        their own snodes, while excluding cascaded history from upstream
-        blocks.
-    """
-    fqn_map: dict[str, str] = V.graph.fx_fqn_map
-    log.debug("get_fused_kernel_module_fqn: snodes=%d", len(scheduler_nodes))
-
-    # Pass 1: derive block anchor prefixes from each snode's origin_node.
-    # origin_node is direct (not accumulated), giving clean block identity.
-    # Fallback: when origin_node is absent or a placeholder (not in fqn_map),
-    # walk origins to find the first non-placeholder op in fqn_map.
-    anchor_prefixes: OrderedSet[str] = OrderedSet()
-    for snode in scheduler_nodes:
-        if snode.node is None:
-            continue
-        origin = snode.node.get_origin_node()
-        origin_name = origin.name if origin is not None else None
-        anchor_fqn = fqn_map.get(origin_name) if origin_name else None
-        if anchor_fqn:
-            stack = origin.meta.get("nn_module_stack")
-            prefix = _outermost_prefix(stack) if stack else None
-            if prefix:
-                anchor_prefixes.add(prefix)
-            continue
-
-        # origin_node absent or not in fqn_map.
-        # For placeholders: scan FX consumers (users) — they identify which
-        # block uses this parameter, not the upstream producers in origins.
-        # For None origin_node: fall back to scanning origins.
-        if origin is not None and origin.op == "placeholder":
-            fallback_source = origin.users
-        else:
-            fallback_source = snode.node.origins
-        for fx_node in fallback_source:
-            fallback_fqn = fqn_map.get(fx_node.name)
-            if fallback_fqn:
-                stack = fx_node.meta.get("nn_module_stack")
-                prefix = _outermost_prefix(stack) if stack else None
-                if prefix:
-                    anchor_prefixes.add(prefix)
-                break
-    if not anchor_prefixes:
-        return None
-
-    # Pass 2: walk all origins across every snode (the transitively accumulated
-    # set), look each up in fqn_map, and include only those whose FQN prefix
-    # matches an anchor.  This captures inline ops (e.g. relu, add inlined into
-    # a parent buffer) while rejecting cascaded history from upstream blocks.
-    extern_fqns: OrderedSet[str] = V.graph.fx_extern_fqns
-    module_names: OrderedSet[str] = OrderedSet()
-    for snode in scheduler_nodes:
-        if snode.node is None:
-            continue
-        for fx_node in snode.node.origins:
-            fqn = fqn_map.get(fx_node.name)
-            if not fqn:
-                continue
-            if fqn in extern_fqns:
-                continue
-            if not any(fqn == p or fqn.startswith(p + ".") for p in anchor_prefixes):
-                continue
-            module_names.add(fqn)
-
-    result = " + ".join(f"L.{fqn}" for fqn in module_names) if module_names else None
-    log.debug("get_fused_kernel_module_fqn: result=%s", result)
-    return result
 
 
 def make_graph_view(
